PyCharm/EstimatorProject/requirements.txt
absl-py==0.7.1
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
colorama==0.4.1
contextlib2==0.5.5
cycler==0.10.0
Cython==0.29.6
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lxml==4.3.2
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.6
numpy==1.16.2
object-detection==0.1
opencv-python==4.0.0.21
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
psutil==5.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
slim==0.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.3.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
